requirements.txt
# Legal RAG Assistant — Dependencies
# ====================================
# Core
sentence-transformers>=2.7.0
numpy>=1.24.0
requests>=2.31.0

# Document loading
pypdf>=4.0.0
python-docx>=1.1.0

# Web UI
streamlit>=1.35.0

# REST API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.7.0

# Optional: local LLM for generation (heavy — skip if not needed)
# transformers>=4.40.0
# torch>=2.2.0
# accelerate>=0.29.0

# Testing
pytest>=8.0.0
httpx>=0.27.0          # for FastAPI test client
